backend/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-unixdatetimefield==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
mongoengine==0.24.1
PyJWT==2.4.0
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
